requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
prompt-toolkit==3.0.38
pytz==2023.3
redis==4.5.4
six==1.16.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
gevent==22.10.2
pymysql==1.0.3
pyjwt==2.6.0
py7zr==0.20.5
google-cloud-storage==2.9.0
google-cloud-pubsub==2.16.1